+++ src/test/java/org/influxdb/InfluxDBTest.java
@@ -41,6 +41,8 @@ import java.util.concurrent.atomic.LongAdder;
 import java.util.function.Consumer;
 import java.util.regex.Pattern;
 
+import static org.assertj.core.api.Assertions.assertThat;
+
 /**
  * Test the InfluxDB API.
  *
@@ -722,6 +724,16 @@ public class InfluxDBTest {
 		Assertions.assertEquals(queryResult.getResults().get(0).getSeries().get(0).getValues().get(1).get(0), timeP2);
 		Assertions.assertEquals(queryResult.getResults().get(0).getSeries().get(0).getValues().get(2).get(0), timeP3);
 
+		// WHEN I use the post query
+		queryResult = this.influxDB.query(new Query("SELECT * FROM " + measurement, dbName, true), TimeUnit.NANOSECONDS);
+
+		// THEN result will be same
+		Assertions.assertEquals(queryResult.getResults().get(0).getSeries().get(0).getValues().size(), 3);
+		Assertions.assertEquals(queryResult.getResults().get(0).getSeries().get(0).getValues().size(), 3);
+		Assertions.assertEquals(queryResult.getResults().get(0).getSeries().get(0).getValues().get(0).get(0), timeP1);
+		Assertions.assertEquals(queryResult.getResults().get(0).getSeries().get(0).getValues().get(1).get(0), timeP2);
+		Assertions.assertEquals(queryResult.getResults().get(0).getSeries().get(0).getValues().get(2).get(0), timeP3);
+		
 		this.influxDB.query(new Query("DROP DATABASE " + dbName));
 	}
 
@@ -1301,6 +1313,37 @@ public class InfluxDBTest {
 
   }
 
+	@Test
+    public void testChunkingQueryPost() throws InterruptedException {
+        if (this.influxDB.version().startsWith("0.") || this.influxDB.version().startsWith("1.0")) {
+            // do not test version 0.13 and 1.0
+            return;
+        }
+
+        String dbName = "write_unittest_" + System.currentTimeMillis();
+        this.influxDB.query(new Query("CREATE DATABASE " + dbName));
+        String rp = TestUtils.defaultRetentionPolicy(this.influxDB.version());
+        BatchPoints batchPoints = BatchPoints.database(dbName).retentionPolicy(rp).build();
+        Point point1 = Point.measurement("disk").tag("atag", "a").addField("used", 60L).addField("free", 1L).build();
+        Point point2 = Point.measurement("disk").tag("atag", "b").addField("used", 70L).addField("free", 2L).build();
+        Point point3 = Point.measurement("disk").tag("atag", "c").addField("used", 80L).addField("free", 3L).build();
+        batchPoints.point(point1);
+        batchPoints.point(point2);
+        batchPoints.point(point3);
+        this.influxDB.write(batchPoints);
+
+        CountDownLatch countDownLatch = new CountDownLatch(2);
+
+        Thread.sleep(2000);
+        Query query = new Query("SELECT * FROM disk", dbName, true);
+        this.influxDB.query(query, 2, result -> countDownLatch.countDown());
+
+        boolean await = countDownLatch.await(10, TimeUnit.SECONDS);
+        Assertions.assertTrue(await, "The QueryResults did not arrive!");
+
+        this.influxDB.query(new Query("DROP DATABASE " + dbName));
+    }
+
   @Test
     public void testFlushPendingWritesWhenBatchingEnabled() {
         String dbName = "flush_tests_" + System.currentTimeMillis();
@@ -1414,6 +1457,18 @@ public class InfluxDBTest {
     Assertions.assertTrue(MyInfluxDBBean.OK);
     //assert that MyInfluxDBBean.OKHTTP_BUILDER stays untouched (no interceptor added)
     Assertions.assertTrue(MyInfluxDBBean.OKHTTP_BUILDER.interceptors().isEmpty());
+  }
+
+  @Test
+  public void testQueryPostWithGZIPCompression() {
+    this.influxDB.enableGzip();
+    String database = "db_gzip_" + System.currentTimeMillis();
+    this.influxDB.query(new Query(String.format("CREATE DATABASE %s", database), null, true));
+    QueryResult query = this.influxDB.query(new Query("SHOW DATABASES", null, true));
+    assertThat(query.getResults()).hasSize(1);
+    assertThat(query.getResults().get(0).getSeries()).hasSize(1);
+    assertThat(query.getResults().get(0).getSeries().get(0).getValues()).contains(Collections.singletonList(database));
+    this.influxDB.query(new Query(String.format("DROP DATABASE %s", database), null, true));
   }
 
   private static final class MyInfluxDBBean {